# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface SocialLinks {
  github?: string;
  linkedin?: string;
  twitter?: string;
  website?: string;
  email?: string;
}

export interface User {
  id: string;
  username: string;
  name: string;
  email: string;
  bio: string;
  headline: string;
  location: string;
  avatarUrl: string;
  resumeUrl: string;
  templateId: string;
  socialLinks: SocialLinks;
  isPro: boolean;
  customDomain: string;
  createdAt: string;
  updatedAt: string;
}

export interface MeResponse {
  user: User | null;
  needsOnboarding: boolean;
}

export interface OnboardInput {
  /**
   * @minLength 3
   * @maxLength 30
   * @pattern ^[a-z0-9_-]+$
   */
  username: string;
  /** @minLength 1 */
  name: string;
  email: string;
}

export interface UpdateUserInput {
  name?: string;
  bio?: string;
  headline?: string;
  location?: string;
  avatarUrl?: string;
  resumeUrl?: string;
  templateId?: string;
  socialLinks?: SocialLinks;
  customDomain?: string;
  isPro?: boolean;
}

export interface UsernameAvailability {
  username: string;
  available: boolean;
}

export interface UserStats {
  projectsCount: number;
  skillsCount: number;
  /** Profile completion percent (0-100) */
  completion: number;
  publicUrl: string;
}

export interface Project {
  id: string;
  userId: string;
  title: string;
  description: string;
  techStack: string[];
  githubLink: string;
  liveLink: string;
  imageUrl: string;
  position: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  techStack?: string[];
  githubLink?: string;
  liveLink?: string;
  imageUrl?: string;
}

export interface UpdateProjectInput {
  title?: string;
  description?: string;
  techStack?: string[];
  githubLink?: string;
  liveLink?: string;
  imageUrl?: string;
  position?: number;
}

export interface Skill {
  id: string;
  userId: string;
  category: string;
  items: string[];
  position: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSkillInput {
  /** @minLength 1 */
  category: string;
  items?: string[];
}

export interface UpdateSkillInput {
  category?: string;
  items?: string[];
  position?: number;
}

export interface PublicProfile {
  user: User;
  projects: Project[];
  skills: Skill[];
}

export interface ShowcaseUser {
  username: string;
  name: string;
  headline: string;
  templateId: string;
  avatarUrl: string;
  isPro: boolean;
}

export interface Template {
  id: string;
  name: string;
  description: string;
  previewBackground: string;
  accentColor: string;
  isPro: boolean;
}

export interface ContactMessageInput {
  username: string;
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactAck {
  accepted: boolean;
}

/**
 * Unauthorized
 */
export type UnauthorizedResponse = Error;

/**
 * Bad request
 */
export type BadRequestResponse = Error;

/**
 * Not found
 */
export type NotFoundResponse = Error;