# lib/api-zod/src/generated/types/contactMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessageInput {
  username: string;
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
}