# lib/api-zod/src/generated/types/createProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  techStack?: string[];
  githubLink?: string;
  liveLink?: string;
  imageUrl?: string;
}