# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: string;
  userId: string;
  category: string;
  items: string[];
  position: number;
  createdAt: Date;
  updatedAt: Date;
}