# lib/api-zod/src/generated/types/updateSkillInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSkillInput {
  category?: string;
  items?: string[];
  position?: number;
}