# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeFolio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialLinks } from "./socialLinks";

export interface User {
  id: string;
  username: string;
  name: string;
  email: string;
  bio: string;
  headline: string;
  location: string;
  avatarUrl: string;
  resumeUrl: string;
  templateId: string;
  socialLinks: SocialLinks;
  isPro: boolean;
  customDomain: string;
  createdAt: Date;
  updatedAt: Date;
}